Conditional Rendering of Login
Added Conditional Rendering of the login state so that if the user is logged in, instead of seeing 'log in' in the header they can be directed to the profile page instead. I've also added a logout button in the profile page that logs the user out and redirects them to the main page.