Lots of work done
As said in the title, lots of work has been done since the last push. I've added a working logout button (with confirmation of logging out) that will display to the user once the button has been pressed, if they say yes they will be navigated back to the main screen.  I added some small functionality to the PostChit button with regards to me either displaying it constantly or only displaying if the user is logged in (The later will need a reload of the main page in order to work correctly). Otherwise they will be presented with an alert asking them to sign in.  On the home page I added a reload button which will display the post button if I want it to, and expanded the information on the FlatList so it shows their userID (More so for a debugging purpose - will probably change this at a later date.) Did a little styling on the LoginPage, created a 'Your Profile' page which has it's own stack nav, (idea is to combine this with tab navs so they can navigate through pages to see their Followers, Following, and maybe put the search user function in there. Finally added the UserProfilePage, this is accessed via pressing on any of the chits in the Main Screen, it will take them to a profile of the person / ID that posted the chit and I'm currently working on displaying their recent chits. Once done I'll work on the photo, or the 'Your Profile' page.